==> notebook/1. preprocess.ipynb <==
import os
import re
import time
import pandas as pd
import numpy as np
import random
import json

from bs4 import BeautifulSoup

def text_strip(_str):
    pattern_strip = re.compile('[\t|\n|\s]*')
    return re.sub(pattern_strip,'',_str)

# 读取权利要求 总结数据

claim_df_filename = '../data/processed/claim_df_0619.pkl'
claim_dir = '../data/raw_data/claim_info/'
if not os.path.isfile(claim_df_filename):
    claim_files = []
    for _dir,_,file in os.walk(claim_dir):
        claim_files+= [os.path.join(_dir,i) for i in file]
    claim_df = pd.DataFrame()
    claim_df['id'] = [i for i in range(len(claim_files))]
    claim_df['pub_id'] = [i.split('/')[-1].split('.')[0] for i in claim_files]
    claim_df['info_id'] = [i.split('/')[-2] for i in claim_files]
    # claim_df['text'] = [text_strip(BeautifulSoup(open(i).read(),'lxml').text) for i in claim_files]
    claim_text = []
    for num,i in enumerate(claim_files):
        if num%10000==0:
            print('[{0}] Num:{1}'.format(time.ctime(),num))
        claim_text.append(text_strip(BeautifulSoup(open(i).read(),'lxml').text))
    claim_df['text'] = claim_text
    claim_df.to_pickle(claim_df_filename)
else:
    claim_df = pd.read_pickle(claim_df_filename)

# 读取专利详情 总结数据

detail_df_filename = '../data/processed/detail_df_0619.pkl'
detail_dir = '../data/raw_data/detail_info/'
if not os.path.isfile(detail_df_filename):
    detail_files = []
    for _dir,_,file in os.walk(detail_dir):
        detail_files += [os.path.join(_dir,i) for i in file]
    detail_df = pd.DataFrame()
    detail_df['id'] = [i for i in range(len(detail_files))]
    detail_df['pub_id'] = [i.split('/')[-1].split('.')[0] for i in detail_files]
    detail_df['info_id'] = [i.split('/')[-2] for i in detail_files]
    detail_data = []
    for num,fname in enumerate(detail_files):
        if num%10000==0:
            print('[{0}] Num:{1}'.format(time.ctime(),num))
        temp_json = json.loads(open(fname).read())
        detail_data.append(temp_json['data'])
    detail_df['data'] = detail_data
    detail_df.to_pickle(detail_df_filename)
else:
    detail_df = pd.read_pickle(detail_df_filename)

# 读取法律信息 总结数据

legal_df_filename = '../data/processed/legal_df_0619.pkl'
legal_dir = '../data/raw_data/legal_info/'

if not os.path.isfile(legal_df_filename):
    legal_files = []
    for _dir,_,file in os.walk(legal_dir):
        legal_files+=[os.path.join(_dir,i) for i in file]
    legal_df = pd.DataFrame()
    legal_df['id'] = [i.split('/')[-1].split('.')[0] for i in legal_files]
    legal_df['info_id'] = [i.split('/')[-2] for i in legal_files]
    legal_df['data'] = [json.loads(open(i).read()) for i in legal_files]
    legal_df.to_pickle(legal_df_filename)
else:
    legal_df = pd.read_pickle(legal_df_filename)
