==> notebook/4.cluster.ipynb <==
import os
import time
import random
import re
import pandas as pd
import numpy as np

import jieba

import pickle

def dump_pickle(obj,fname):
    with open(fname,'wb') as file:
        pickle.dump(obj,file)
def load_pickle(fname):
    pickle.load(open(fname,'rb').read())

claim_df = pd.read_pickle('../data/processed/claim_df_0619.pkl')

claim_df

# 预处理一下

1、数据清洗（把里面的0001之类的删掉）

2、分词（专利法律词典）


patent_word = [i.strip() for i in open('../data/user_data/patent_word_list.txt','r',encoding='utf-8').readlines()]

patent_word = [i.split(' ')[-1] for i in patent_word]

for word in patent_word:
    jieba.add_word(word)

def wash_number(_str):
    pattern_num = re.compile('\d{4}\.')
    pattern_num1 = re.compile('\d{1}\.')
    _str = re.sub(pattern_num,'',_str)
    _str = re.sub(pattern_num1,'',_str)
    return _str

claim_df['text'] = claim_df.text.apply(lambda x:wash_number(x))

claim_df

claim_df['seg_text'] = claim_df.text.apply(jieba.lcut)

claim_df.to_pickle('../data/processed/claim_df_seg_0630.pkl')

# 向量化 使用tfidf

from sklearn.feature_extraction.text import TfidfVectorizer

patent_text = claim_df.seg_text.tolist()
patent_text = [' '.join(i) for i in patent_text]

patent_tfidf_obj = TfidfVectorizer()

patent_tfidf_obj.fit(patent_text)

dump_pickle(patent_tfidf_obj,'../data/processed/patent_tfidf_obj_default_0701.pkl')

patent_all_tfidf_matrix = patent_tfidf_obj.transform(patent_text)

# PCA 降维对象

from sklearn.decomposition import TruncatedSVD

pca_obj = TruncatedSVD(n_components=250)

%timeit
pca_obj.fit(patent_all_tfidf_matrix)

dump_pickle(pca_obj,'../data/processed/patent_pca_obj_default_0701.pkl')

# LDA 降维对象

from sklearn.decomposition import LatentDirichletAllocation

lda_obj = LatentDirichletAllocation(n_components=100,)

lda_obj.fit(patent_all_tfidf_matrix)

dump_pickle(lda_obj,'../data/processed/patent_lda_obj_default_0701.pkl')

# 数据准备

sample_patent_data = patent_tfidf_obj.transform(patent_text[:10000])

# 聚类 使用 k-means

from sklearn.cluster import k_means

km = k_means(sample_patent_data,5)

km[1]

# 聚类 使用 AffinityPropagation

自动确定团簇个数

predict 方法能够输入新文档后确定所在团簇

cluster_centers_ 给出团簇中心

from sklearn.cluster import AffinityPropagation

afp = AffinityPropagation()

afp.fit(sample_patent_data)

afp.predict(sample_patent_data[:5])

afp.cluster_centers_

# 聚类 使用

from sklearn.cl
