==> notebook/21.paper sql data manage.ipynb <==
import os
import time
import random
import re
import pandas as pd
import numpy as np
import mysql.connector

host = '127.0.0.1'
db_user = 'bio_user'
db_password = 'ir@DUT#bio'
db_name = 'bioDB'

## 新建数据库

# cnx = mysql.connector.connect(user='root', password='ir@DUT#2018',
#                               host='127.0.0.1')

# cursor = cnx.cursor()

# cursor.execute('show databases;')

# cursor.execute('create database patentDB;')

# cursor.execute('show databases;')

# cnx.close()

## 新建paper数据表

cnx = mysql.connector.connect(user=db_user, password=db_password,
                              host=host,database=db_name)

cursor = cnx.cursor()

cursor.execute('drop table paper_table;')

cursor.execute('create table paper_table \
               (P_ID int not null primary key auto_increment,\
               P_Title text,P_Author text, P_Publication text,P_Organ text, P_year int, P_Keyword text,\
               P_Summary text, P_Keyword_seg text, P_Title_seg text, P_Summary_seg text, P_URL text, P_Fields text, P_Fields_two text, P_References text,\
               P_Pagecount int, P_Page text, P_Language text, P_Download_num int,P_Citation_num int, P_Vector text, P_Volume text,\
               P_Issue text, P_Issn text, P_Isbn text, P_Doi text, P_Red1 text, P_Red2 text, P_Red3 text, P_Red4 text, P_Red5 text);')
cnx.close()

## 新建author数据表

cnx = mysql.connector.connect(user=db_user, password=db_password,
                              host=host,database=db_name)

cursor = cnx.cursor()

author_table_str = 'create table author_table\
(A_ID int not null primary key auto_increment,\
    A_Cname text, A_Ename text, A_Normalized_name text,\
    A_Organ text, A_Position text, A_Fields text, A_Pubs_num int,\
    A_Citation_num int, A_Index int, A_Research_interest text, A_Focus_Fields text,\
    A_Paper_ID text, A_Co_ID text, A_Co_name text, A_Vector text, A_Red1 text, A_Red2 text, A_Red3 text, A_Red4 text, A_Red5 text);'

cursor.execute(author_table_str)

cnx.close()

## 新建Year数据建表

cnx = mysql.connector.connect(user=db_user, password=db_password,
                              host=host,database=db_name)

cursor = cnx.cursor()

cursor.execute('drop table year_table;')

year_table_str = 'create table year_table \
(Y_ID int not null primary key auto_increment,Y_Year int,Y_Title_seg text, Y_Keyword text, Y_Summary_seg text, Y_Authors text,\
    Y_Publication text, Y_Number int, Y_Fields text,Y_Fields_two text, Y_Red1 text, Y_Red2 text, Y_Red3 text);'

cursor.execute(year_table_str)

cnx.close()
