==> notebook/31. ES paper-index.ipynb <==
import time
import random
import os
import re
import pandas as pd
import numpy as np
from datetime import datetime
from elasticsearch import Elasticsearch 
import mysql.connector
from elasticsearch.helpers import bulk

# 日期处理

def date_convert(dot_date_str):
    '''xxxx.xx.xx --> xxxx-xx-xx
    then pack into json for es
    '''
    y,m,d = dot_date_str.split('.')
    new_date = '{year}-{month}-{date}'.format(year=y,month=m,date=d)
    return new_date
#     return {new_date:{'type':'date','formate':'yyyy-mm-dd'}}

def date_autofill(dot_date_str):
    '''xxxx --> xxxx-01-01
    then pack into json for es
    '''
    new_date = '{year}-{month}-{date}'.format(year=dot_date_str,month='01',date='01')
    return new_date

# 读取库中paper数据

# head = 'P_Title, P_Author, P_Publication, P_Organ, P_year, P_Keyword, \
#     P_Summary, P_Keyword_seg, P_Title_seg,\
#     P_Summary_seg, P_URL, P_Fields, P_Fields_two,P_References, P_Pagecount, P_Page, P_Language,\
#     P_Download_num, P_Citation_num,P_Vector,P_Volume, P_Issue,P_Issn,P_Isbn, P_Doi,\
#     P_Red1, P_Red2, P_Red3, P_Red4, P_Red5'

# head_columns = [i.strip() for i in head.split(',')]

cnx = mysql.connector.connect(user='root', password='ir@DUT#2018',
                              host='10.8.128.205',port='23306')

cursor = cnx.cursor()

# 获取paper表头

cursor.execute("select * from information_schema.columns where table_name='paper_table'")

sql_table_columns = cursor.fetchall()

sql_table_columns_name = [i[3] for i in sql_table_columns]

# 获取paper表总容量

cursor.execute("select table_rows from information_schema.tables where table_name='paper_table'")

sql_table_rows = cursor.fetchall()

sql_table_rows = sql_table_rows[0][0]

def data_to_json(paper_atom,head_columns):
    data = paper_atom
    data_dict = {k:data[num] for num,k in enumerate(head_columns)}
    data_dict['P_year'] = date_autofill(data_dict['P_year'])
    return data_dict

# 构建索引

es = Elasticsearch(hosts='10.8.128.205',port=49200,timeout=30, max_retries=10, retry_on_timeout=True)

# 每次查询limit数目的数值 同时批量插入paper

es.indices.delete(index='paper-year-index')

index_name = 'paper-detail-index'
limit = 5000
offset = 0
for epoch in range(int(sql_table_rows/limit)):
    print('[{0}] Epoch:{1}'.format(time.ctime(),epoch))
    base_fetch_sql = 'select * from {dbName}.{tableName} limit {limitVolume} offset {offsetVolume}'
    base_fetch_sql = base_fetch_sql.format(dbName='bioDB',tableName='paper_table',limitVolume=limit,offsetVolume=offset)
    cursor.execute(base_fetch_sql)
    sql_data = cursor.fetchall()
    sql_dict = [data_to_json(i,sql_table_columns_name) for i in sql_data]
    bulk(es,sql_dict,index=index_name,raise_on_error=True)
    offset += limit - 1
# 最后一桶
base_fetch_sql = 'select * from {dbName}.{tableName} limit {limitVolume} offset {offsetVolume}'
base_fetch_sql = base_fetch_sql.format(dbName='bioDB',tableName='paper_table',limitVolume=limit,offsetVolume=offset)
cursor.execute(base_fetch_sql)
sql_data = cursor.fetchall()
sql_dict = [data_to_json(i,sql_table_columns_name) for i in sql_data]
bulk(es,sql_dict,index=index_name,raise_on_error=True)

cursor.close()

sql_table_rows

# 读取库中author数据

head = 'A_Cname, A_Ename, A_Normalized_name, A_Organ, A_Position, A_Fields , A_Pubs_num,\
    A_Citation_num, A_Index, A_Research_interest, A_Focus_Fields, A_Paper_ID, A_Co_ID, A_Co_name, A_Vector,\
    A_Red1, A_Red2, A_Red3, A_Red4, A_Red5'

head_columns = [i.strip() for i in head.split(',')]

cnx = mysql.connector.connect(user='root', password='ir@DUT#2018',
                              host='10.8.128.205',port='23306')

cursor = cnx.cursor()

# 获取paper表头

cursor.execute("select * from information_schema.columns where table_name='author_table'")

sql_table_columns = cursor.fetchall()

sql_table_columns_name = [i[3] for i in sql_table_columns]

sql_table_columns_name

# 获取paper表总容量

cursor.execute("select table_rows from information_schema.tables where table_name='author_table'")

sql_table_rows = cursor.fetchall()

sql_table_rows = sql_table_rows[0][0]

def data_to_json_author(data_atom):
    data = data_atom[1:]
    data_dict = {k:data[num] for num,k in enumerate(head_columns)}
    return data_dict

# 每次查询limit数目的数值 同时批量插入author

from elasticsearch.helpers import bulk

es = Elasticsearch(hosts='10.8.128.205',port=49200,timeout=30, max_retries=10, retry_on_timeout=True)

limit = 10000
offset = 0
for epoch in range(int(sql_table_rows/limit)):
    print('[{0}] Epoch:{1}'.format(time.ctime(),epoch))
    base_fetch_sql = 'select * from {dbName}.{tableName} limit {limitVolume} offset {offsetVolume}'
    base_fetch_sql = base_fetch_sql.format(dbName='bioDB',tableName='author_table',limitVolume=limit,offsetVolume=offset)
    cursor.execute(base_fetch_sql)
    sql_data = cursor.fetchall()
    sql_dict = [data_to_json_author(i) for i in sql_data]
    bulk(es,sql_dict,index='paper-author-index',raise_on_error=True)
    offset += limit - 1
# 最后一桶
base_fetch_sql = 'select * from {dbName}.{tableName} limit {limitVolume} offset {offsetVolume}'
base_fetch_sql = base_fetch_sql.format(dbName='bioDB',tableName='author_table',limitVolume=limit,offsetVolume=offset)
cursor.execute(base_fetch_sql)
sql_data = cursor.fetchall()
sql_dict = [data_to_json_author(i) for i in sql_data]
bulk(es,sql_dict,index='paper-author-index',raise_on_error=True)

# 设置ES状态为可写（触发硬盘存储超限后集群会自动配置所有索引为只读）

import json

import requests

headers = {
    'Accept': 'application/json, text/javascript, */*; q=0.01',
    'Content-Type': 'application/json',
    'Origin': 'chrome-extension://ffmkiejjmecolpfloofpjologoblkegm',
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.100 Safari/537.36'
}

unlock_dict = {
"index.blocks.read_only_allow_delete": "false"
}

ppp = requests.put('http://10.8.128.205:49200/paper-year-index/_settings',data=json.dumps(unlock_dict),headers=headers)

ppp.text

# 论文检索对象

ess = esPatentRetrieval(host='10.8.128.205',port=29200)

ess.Retrieval('机器人','2000-01-22','2010-01-22',volume=100)
