==> notebook/32. ES weibo-index.ipynb <==
import time
import random
import os
import re
import pandas as pd
import numpy as np
from datetime import datetime
from elasticsearch import Elasticsearch 

import mysql.connector

import json

# 日期处理

def date_convert(dot_date_str):
    '''xxxx.xx.xx --> xxxx-xx-xx
    then pack into json for es
    '''
    y,m,d = dot_date_str.split('.')
    new_date = '{year}-{month}-{date}'.format(year=y,month=m,date=d)
    return new_date
#     return {new_date:{'type':'date','formate':'yyyy-mm-dd'}}

def date_autofill(dot_date_str):
    '''xxxx --> xxxx-01-01
    then pack into json for es
    '''
    new_date = '{year}-{month}-{date}'.format(year=dot_date_str,month='01',date='01')
    return new_date

def data_to_json(data_atom,columns):
    data = data_atom
    data_dict = {k:str(data[num]) for num,k in enumerate(columns)}
    return data_dict

# 读取weibo数据

comment_df = pd.read_pickle('../data/weibo_data/comment.pkl')
weibo_df = pd.read_pickle('../data/weibo_data/weibo.pkl')

weibo_df

comment_columns = comment_df.columns.tolist()

comment_columns = ['w_id'] + comment_columns[1:]

comment_columns

weibo_columns = weibo_df.columns.tolist()

weibo_columns = ['w_id'] + weibo_columns[1:]

# 创建ES对象

from elasticsearch.helpers import bulk

from elasticsearch.helpers import bulk

es = Elasticsearch(hosts='10.8.128.205',port=49200,timeout=30, max_retries=10, retry_on_timeout=True)

# 批量插入comment

sample_item = comment_df.iloc[num]
sample_dict = data_to_json(sample_item,comment_columns)
es.create(index='weibo-comment-index',id='test',body=json.dumps(sample_dict))

counter = 0
index_batch = []
batch = 10000
for num,_ in enumerate(comment_df.itertuples()):
    if num %batch==0:
        print('[{0}] Num:{1}'.format(time.ctime(),num))
    sample_item = comment_df.iloc[num]
    sample_dict = data_to_json(sample_item,comment_columns)
    if counter <batch:
        index_batch.append(sample_dict)
        counter += 1
    else:
        index_batch.append(sample_dict)
        print('Bulking....')
        time1 = time.time()
        bulk(es,index_batch,index='weibo-comment-index',raise_on_error=True)
        time2 = time.time()
        print('Total time:{0:.2f}min'.format((time2-time1)/60))
        counter = 0
        index_batch = []
bulk(es,index_batch,index='weibo-comment-index',raise_on_error=True)
es.indices.refresh(index="weibo-comment-index")

# 批量插入weibo

sample_item = weibo_df.iloc[0]
sample_dict = data_to_json(sample_item,weibo_columns)

es.create(index='weibo-tweet-index',id='test',body=json.dumps(sample_dict))

counter = 0
index_batch = []
batch = 10000
for num,_ in enumerate(weibo_df.itertuples()):
    if num %batch==0:
        print('[{0}] Num:{1}'.format(time.ctime(),num))
    sample_item = weibo_df.iloc[num]
    sample_dict = data_to_json(sample_item,comment_columns)
    if counter <batch:
        index_batch.append(sample_dict)
        counter += 1
    else:
        index_batch.append(sample_dict)
        print('Bulking....')
        time1 = time.time()
        bulk(es,index_batch,index='weibo-tweet-index',raise_on_error=True)
        time2 = time.time()
        print('Total time:{0:.2f}min'.format((time2-time1)/60))
        counter = 0
        index_batch = []
bulk(es,index_batch,index='weibo-tweet-index',raise_on_error=True)
es.indices.refresh(index="weibo-tweet-index")
