==> notebook/42.text cluster test (tfidf+kmeans).ipynb <==
import os
import time
import random
import re
import pandas as pd
import numpy as np

import jieba

import pickle

def dump_pickle(obj,fname):
    with open(fname,'wb') as file:
        pickle.dump(obj,file)
def load_pickle(fname):
    pickle.load(open(fname,'rb').read())

claim_df = pd.read_pickle('../data/patent_data/processed/claim_df_seg_0630.pkl')

claim_df

# 向量化 使用tfidf 限定1w个词

from sklearn.feature_extraction.text import TfidfVectorizer

patent_text = claim_df.seg_text.tolist()
patent_text = [' '.join(i) for i in patent_text]

patent_tfidf_obj = TfidfVectorizer(max_features=10000)

patent_tfidf_obj.fit(patent_text)

dump_pickle(patent_tfidf_obj,'../data/patent_data/model/patent_tfidf_model.pkl')

patent_all_tfidf_matrix = patent_tfidf_obj.transform(patent_text)

# PCA 降维对象

from sklearn.decomposition import TruncatedSVD

pca_obj = TruncatedSVD(n_components=250)

pca_obj.fit(patent_all_tfidf_matrix)

# LDA 降维对象

from sklearn.decomposition import LatentDirichletAllocation

baseNmaeTemplate = '../data/patent_data/model/patent_lda_topic{0}_model.pkl'

for i in range(2,10):
    print('Model LDA topic:',i)
    lda_obj = LatentDirichletAllocation(n_components=i,)
    lda_obj.fit(patent_all_tfidf_matrix)
    dump_pickle(lda_obj,baseNmaeTemplate.format(i))

lda_obj = LatentDirichletAllocation(n_components=100,n_jobs=12)
lda_obj.fit(patent_all_tfidf_matrix)
dump_pickle(lda_obj,baseNmaeTemplate.format(100))

# 数据准备

sample_patent_data = patent_tfidf_obj.transform(patent_text[:10000])

sample_pca_data = pca_obj.transform(sample_patent_data)

sample_lda_data = lda_obj.transform(sample_patent_data)

# 聚类 使用 k-means

from sklearn.cluster import KMeans

km_tfidf = KMeans(n_clusters=8)

km_pca = KMeans(n_clusters=8)

km_lda = KMeans(n_clusters=8)

cluster_tfidf = km_tfidf.fit_predict(sample_patent_data)

cluster_pca = km_pca.fit_predict(sample_pca_data)

cluster_lda = km_lda.fit_predict(sample_lda_data)

# 词云输出关键词

claim_kwrod_tfidf_df = pd.read_pickle('../data/patent_data/processed/claim_kword_df_tfidf.pkl')

claim_kword_textrank_df = pd.read_pickle('../data/patent_data/processed/claim_kword_df_tr4w.pkl')
