==> notebook/52.weibo retrieval comment retireval.ipynb <==
# 将检索模块和文本聚类模块进行封装

import time
import random
import os
import re
import pandas as pd
import numpy as np
from datetime import datetime
from elasticsearch import Elasticsearch

import mysql.connector

# ES检索模块重构为一个对象

columns = ['C_ID',
 'C_comment_id',
 'C_comment_user_id',
 'C_comment_user_nick_name',
 'C_content',
 'C_weibo_url',
 'C_like_num',
 'C_created_at',
 'C_crawl_time']

','.join(columns)

class esWeiboCommentRetrieval():
    '''
    根据论文检索需求进行功能的微调
    '''
    _instance = None
    _first_init = True
    def __new__(cls, *args, **kw):
        if not cls._instance:
            cls._instance = super(esWeiboCommentRetrieval, cls).__new__(cls)  
        return cls._instance
    
    def __init__(self, host, port):
        '''
        使用ES进行论文检索 指定host、port以及专利index之后进行检索
        '''
        super(esWeiboCommentRetrieval, self).__init__()
        self.es = Elasticsearch(hosts=host, port=port, timeout=30, max_retries=10, retry_on_timeout=True)
        self.indexName = 'weibo-comment-index'

    def do_search(self, wordQuery, textQuery, volume):
        '''
        do_search方法执行具体检索过程
        wordQuery 本应为查询对应微博时所用的检索词 此处暂时不用 目前暂时只检索微博评论数据后续再修改对象为
        
        volume为每次检索返回的数目
        '''
        queryBody = {
          "query": {
            "match": {
              "C_content": textQuery
            }
          },
          "from": 0,
          "size": volume,
          "sort": [],
          "aggs": {}
        }
        result = self.es.search(index=self.indexName, body=queryBody)
        return result

    def format_search(self, result):
        '''
        format_search方法对检索结果进行格式化 构建符合要求的字段进行返回
        输入result为检索结果 提取其中的检索结果进行后处理
        使用ES检索后得到的结果中result['hits']['hits']为数组格式数据
        其中每一个元素为一个dict 对应部分字段
        '''
        docs = result['hits']['hits']
        docs = [i['_source'] for i in docs]
        targetKeyList = 'C_ID,C_comment_id,C_comment_user_id,C_comment_user_nick_name,C_content,\
        C_weibo_url,C_like_num,C_created_at,C_crawl_time'
        targetKeyList = [i.strip() for i in targetKeyList.split(',')]
        dict_filter_by_keys = lambda d: {k: d[k] for k in targetKeyList}
        dict_filter_text = lambda d: {k if not k == 'text' else 'claim_text': d[k] for k in d}
        dict_filter_id = lambda d: {k if not k == '_id' else 'id': d[k] for k in d}
        docs = (dict_filter_by_keys(doc) for doc in docs)
        docs = [dict_filter_text(doc) for doc in docs]
        return docs
    def Retrieval(self, wordQuery, textQuery, volume):
        result = self.do_search(wordQuery, textQuery, volume)
        docs = self.format_search(result)
        return docs

eee = esWeiboCommentRetrieval(host='10.8.128.205',port=49200,)

result= eee.Retrieval( wordQuery='我', textQuery='我',volume=100)

# 客户端对于后端数据接口的访问

import requests

body = {'wordQuery':'我',
        'textQuery':'垃圾分类',
        'volume':1000}

baseUrl = 'http://10.8.128.205:29280/Lawbda/dataWare/1.0.0/weibo/search'

docs = requests.get(baseUrl,params=body)

docs.json()[0]

def get_hot_topic_and_weibos():
    """
    获取热门话题和话题对应的热门微博，每个话题封装成一个字典
    :return:  字典列表，[one_topic_dct{topic, content, nick_name, ...}]
    """
    coon = mysql.connector.connect(user='weibo_user', password='ir@DUT#weibo', host="10.8.128.205",
                                   port='23306')
    cursor = coon.cursor(buffered=True)
    cursor.execute('use weiboDB;')

    ##  热门话题，这里先给定，以后完善
    hot_topics = ['#垃圾分类#', '#韩商言借钱#', '#被辱骂初一女生家长原谅班主任#', '#赵丽颖产后现身#', '#李荣浩成功求婚杨丞琳#',
                  '#2018结婚率创十年新低#', '#中国花协推荐牡丹为国花#', '#2018结婚率创十年新低#', '#香港警方拘捕47人#', '#精神病患是否可坐飞机#', ]

    hot_topics_msg_lst = []
    for hot_topic in hot_topics:
        # 挑选每个热门话题的热门微博，这里只选择评论最多的一条
        # 从数据库中根据 W_search_query 的值查找，返回5列内容
        sql_base = 'select W_content, W_nick_name, W_repost_num, W_comment_num, W_like_num \
                   from weibo_table where W_search_query="{}";'
        sql_str = sql_base.format(hot_topic)
        cursor.execute(sql_str)
        items = cursor.fetchall()

        hot_weibo = sorted(items, key=lambda x: x[3], reverse=True)[0]
        hot_weibo_dct = {'topic': hot_topic,
                         'content': hot_weibo[0], 'nick_name': hot_weibo[1],
                         'repost_num': hot_weibo[2], 'comment_num': hot_weibo[3],
                         'like_num': hot_weibo[4]}
        hot_topics_msg_lst.append(hot_weibo_dct)

    coon.close()
    return hot_topics_msg_lst

get_hot_topic_and_weibos()
